demo/requirements.txt
beautifulsoup4==4.12.0
bs4==0.0.1
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.3
img2table==0.0.19
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pdf2image==1.16.3
pdfextract==0.1.2
Pillow==9.4.0
polars==0.16.14
pyarrow==11.0.0
PyMuPDF==1.21.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
soupsieve==2.4
typing_extensions==4.5.0
Werkzeug==2.2.3
XlsxWriter==3.0.9
zipp==3.15.0
